Extract common code for Payment Sources from Credit Card
