+++ core/spec/models/spree/payment_source_spec.rb
@@ -1,0 +1,9 @@
+require 'spec_helper'
+
+describe Spree::PaymentSource, type: :model do
+  it 'is an abstract class' do
+    expect do
+      described_class.new
+    end.to raise_error(NotImplementedError)
+  end
+end
